Usage of package objects avoided.
