--- src/main/scala/encrywm/lib/predef/functions/signature/package.scala
@@ -1,10 +1,0 @@
-package encrywm.lib.predef.functions
-
-import encrywm.lang.backend.env.ESEnvComponent
-
-package object signature {
-
-  val predefFunctions: Map[String, ESEnvComponent] = Map(
-    CheckSig.name -> CheckSig.asFunc
-  )
-}
