--- src/main/scala/encrywm/lib/predef/functions/time/package.scala
@@ -1,10 +1,0 @@
-package encrywm.lib.predef.functions
-
-import encrywm.lang.backend.env.ESEnvComponent
-
-package object time {
-
-  val predefFunctions: Map[String, ESEnvComponent] = Map(
-    Str2Time.name -> Str2Time.asFunc
-  )
-}
