+++ src/main/scala/encrywm/lib/predef/hash/Keccak256Hash.scala
@@ -1,14 +1,14 @@
-package encrywm.lib.predef.functions.hash
+package encrywm.lib.predef.hash
 
 import encrywm.lang.backend.env.{ESBuiltInFunc, ESValue}
 import encrywm.lang.backend.executor.error.BuiltInFunctionExecException
 import encrywm.lib.Types.ESByteVector
-import encrywm.lib.predef.functions.BuiltInFunctionHolder
-import scorex.crypto.hash.{Sha256 => ScorexSha256}
+import encrywm.lib.predef.BuiltInFunctionHolder
+import scorex.crypto.hash.Keccak256
 
-object Sha256Hash extends BuiltInFunctionHolder {
+object Keccak256Hash extends BuiltInFunctionHolder {
 
-  val name = "sha256hash"
+  val name = "keccak256hash"
 
   override def asFunc: ESBuiltInFunc = ESBuiltInFunc(name, args, body)
 
@@ -19,7 +19,7 @@ object Sha256Hash extends BuiltInFunctionHolder {
     val validArgTypes = args.forall { case (_, v) => v.tpe.isInstanceOf[ESByteVector.type] }
     if (validNumberOfArgs && validArgTypes) {
       val fnArgs = args.map(_._2.value.asInstanceOf[Array[Byte]])
-      Right(ScorexSha256.hash(fnArgs.head))
+      Right(Keccak256.hash(fnArgs.head))
     } else {
       Left(BuiltInFunctionExecException)
     }
