+++ src/main/scala/encrywm/lib/predef/hash/Keccak512Hash.scala
@@ -1,14 +1,14 @@
-package encrywm.lib.predef.functions.hash
+package encrywm.lib.predef.hash
 
 import encrywm.lang.backend.env.{ESBuiltInFunc, ESValue}
 import encrywm.lang.backend.executor.error.BuiltInFunctionExecException
 import encrywm.lib.Types.ESByteVector
-import encrywm.lib.predef.functions.BuiltInFunctionHolder
-import scorex.crypto.hash.Keccak256
+import encrywm.lib.predef.BuiltInFunctionHolder
+import scorex.crypto.hash.Keccak512
 
-object Keccak256Hash extends BuiltInFunctionHolder {
+object Keccak512Hash extends BuiltInFunctionHolder {
 
-  val name = "keccak256hash"
+  val name = "keccak512hash"
 
   override def asFunc: ESBuiltInFunc = ESBuiltInFunc(name, args, body)
 
@@ -19,7 +19,7 @@ object Keccak256Hash extends BuiltInFunctionHolder {
     val validArgTypes = args.forall { case (_, v) => v.tpe.isInstanceOf[ESByteVector.type] }
     if (validNumberOfArgs && validArgTypes) {
       val fnArgs = args.map(_._2.value.asInstanceOf[Array[Byte]])
-      Right(Keccak256.hash(fnArgs.head))
+      Right(Keccak512.hash(fnArgs.head))
     } else {
       Left(BuiltInFunctionExecException)
     }
