+++ src/main/scala/encrywm/lib/predef/hash/Sha256Hash.scala
@@ -1,25 +1,25 @@
-package encrywm.lib.predef.functions.signature
+package encrywm.lib.predef.hash
 
 import encrywm.lang.backend.env.{ESBuiltInFunc, ESValue}
 import encrywm.lang.backend.executor.error.BuiltInFunctionExecException
 import encrywm.lib.Types.ESByteVector
-import encrywm.lib.predef.functions.BuiltInFunctionHolder
-import scorex.crypto.signatures.{Curve25519, PublicKey, Signature}
+import encrywm.lib.predef.BuiltInFunctionHolder
+import scorex.crypto.hash.{Sha256 => ScorexSha256}
 
-object CheckSig extends BuiltInFunctionHolder {
+object Sha256Hash extends BuiltInFunctionHolder {
 
-  val name: String = "checkSig"
+  val name = "sha256hash"
 
   override def asFunc: ESBuiltInFunc = ESBuiltInFunc(name, args, body)
 
-  val args = IndexedSeq("sig" -> ESByteVector, "msg" -> ESByteVector, "pubKey" -> ESByteVector)
+  val args = IndexedSeq("input" -> ESByteVector)
 
   private val body = (args: Seq[(String, ESValue)]) => {
-    val validNumberOfArgs = args.size == 3
+    val validNumberOfArgs = args.size == 1
     val validArgTypes = args.forall { case (_, v) => v.tpe.isInstanceOf[ESByteVector.type] }
     if (validNumberOfArgs && validArgTypes) {
       val fnArgs = args.map(_._2.value.asInstanceOf[Array[Byte]])
-      Right(Curve25519.verify(Signature @@ fnArgs.head, fnArgs(1), PublicKey @@ fnArgs.last))
+      Right(ScorexSha256.hash(fnArgs.head))
     } else {
       Left(BuiltInFunctionExecException)
     }
